Fix rebuild_database options

+++ rebuild_database.sh
@@ -22,20 +22,30 @@ if [[ $# -eq 2 ]]; then
     rm db.sqlite3
     cp ./0001_initial_${name}.py tourbuilder/migrations/0001_initial.py
     cp ./db_${name}.sqlite3 db.sqlite3
+  else
+    echo "Unrecognised option"
+  fi
+elif [[ $# -eq 1 ]]; then
+  if [[ "${1}" == "--clear" ]]; then
+    echo "Deleting database"
+    rm -rf tourbuilder/migrations/*
+    rm db.sqlite3
+    python3 manage.py makemigrations tourbuilder
+    python3 manage.py migrate
+    echo "from django.contrib.auth import get_user_model; User = get_user_model(); User.objects.create_superuser('${USERNAME}', '${EMAIL_ADDRESS}', '${PASSWORD}')" | python3 manage.py shell
+  elif [[ "${1}" == "--update" ]]; then
+    echo "Updating database"
+    cd -
+    python3 ./generate_database_fixtures.py
+    echo "Built the fixtures, starting to import"
+    python3 ./website/manage.py loaddata django_db_fixtures.json
+    echo "Finised importing fixtures"
+  else
+    echo "Unrecognised option"
   fi
 else
-  echo "Wiping and restoring database - remove the exit statement from this script!"
-  exit 1
-  rm -rf tourbuilder/migrations/*
-  rm db.sqlite3
-
-  python3 manage.py makemigrations tourbuilder
-  python3 manage.py migrate
-  echo "from django.contrib.auth import get_user_model; User = get_user_model(); User.objects.create_superuser('${USERNAME}', '${EMAIL_ADDRESS}', '${PASSWORD}')" | python3 manage.py shell
-
-  cd -
-  python3 ./generate_database_fixtures.py
-  echo "Built the fixtures, starting to import"
-  python3 ./website/manage.py loaddata django_db_fixtures.json
-  echo "Finised importing fixtures"
+  echo "Usage: --save NAME"
+  echo "Usage: --restore NAME"
+  echo "Usage: --clear"
+  echo "Usage: --update"
 fi
